apps/worker/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
redis==5.2.1
